+++ src/main/java/com/epam/reportportal/jbehave/ReportPortalStoryReporter.java
@@ -38,6 +38,7 @@ import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
 import org.jbehave.core.model.*;
 import org.jbehave.core.reporters.NullStoryReporter;
+import org.jbehave.core.steps.Timing;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -73,6 +74,7 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	private static final String EXAMPLE_VALUE_PATTERN = "<%s>";
 	private static final Pattern EXAMPLE_VALUE_MATCH = Pattern.compile("<([^>]*)>");
 	private static final String EXAMPLE = "EXAMPLE";
+	private static final String LIFECYCLE = "LIFECYCLE";
 	private static final String EXAMPLE_PARAMETER_DELIMITER = PARAMETER_ITEMS_DELIMITER + " ";
 	private static final String EXAMPLE_KEY_VALUE_DELIMITER = CODE_REFERENCE_ITEM_TYPE_DELIMITER + " ";
 	private static final String NO_NAME = "No name";
@@ -85,8 +87,10 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	private final Deque<TestItemTree.TestItemLeaf> stepStack = new LinkedList<>();
 	private final Supplier<Launch> launch;
 	private final TestItemTree itemTree;
-	private volatile ItemType currentLifecycleItemType;
+
 	private volatile TestItemTree.TestItemLeaf lastStep;
+	private static volatile ItemType currentLifecycleTopItemType;
+	private volatile ItemType currentLifecycleItemType;
 
 	public ReportPortalStoryReporter(final Supplier<Launch> launchSupplier, TestItemTree testItemTree) {
 		launch = launchSupplier;
@@ -111,15 +115,27 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @param type          an item type
 	 * @return a code reference string to identify every element of a story
 	 */
-	private String getCodeRef(@Nullable final String parentCodeRef, @Nonnull final TestItemTree.ItemTreeKey key, ItemType type) {
+	private String getCodeRef(@Nullable final String parentCodeRef, @Nonnull final TestItemTree.ItemTreeKey key,
+			@Nonnull ItemType type) {
 		StringBuilder sb = new StringBuilder();
 		if (isBlank(parentCodeRef)) {
 			sb.append(key.getName());
 		} else {
+			String typeName;
+			switch (type) {
+				case SUITE:
+					typeName = EXAMPLE;
+					break;
+				case TEST:
+					typeName = LIFECYCLE;
+					break;
+				default:
+					typeName = type.name();
+			}
 			sb.append(parentCodeRef)
 					.append(CODE_REFERENCE_DELIMITER)
 					.append(CODE_REFERENCE_ITEM_START)
-					.append(type != ItemType.SUITE ? type.name() : EXAMPLE)
+					.append(typeName)
 					.append(CODE_REFERENCE_ITEM_TYPE_DELIMITER)
 					.append(key.getName().replace("\n", "").replace("\r", ""))
 					.append(CODE_REFERENCE_ITEM_END);
@@ -173,7 +189,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @return Request to ReportPortal
 	 */
 	@Nonnull
-	protected StartTestItemRQ buildStartStoryRq(@Nonnull Story story, @Nonnull String codeRef, @Nullable final Date startTime) {
+	protected StartTestItemRQ buildStartStoryRq(@Nonnull Story story, @Nonnull String codeRef,
+			@Nullable final Date startTime) {
 		StartTestItemRQ rq = new StartTestItemRQ();
 		rq.setName(getStoryName(story));
 		rq.setCodeRef(codeRef);
@@ -215,7 +232,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @return Request to ReportPortal
 	 */
 	@Nonnull
-	protected StartTestItemRQ buildStartScenarioRq(@Nonnull Scenario scenario, @Nonnull String codeRef, @Nullable final Date startTime) {
+	protected StartTestItemRQ buildStartScenarioRq(@Nonnull Scenario scenario, @Nonnull String codeRef,
+			@Nullable final Date startTime) {
 		StartTestItemRQ rq = new StartTestItemRQ();
 		rq.setName(getScenarioName(scenario));
 		rq.setCodeRef(codeRef);
@@ -236,7 +254,10 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 				example.entrySet()
 						.stream()
 						.map(e -> e.getKey() + EXAMPLE_KEY_VALUE_DELIMITER + e.getValue())
-						.collect(Collectors.joining(EXAMPLE_PARAMETER_DELIMITER, CODE_REFERENCE_ITEM_START, CODE_REFERENCE_ITEM_END))
+						.collect(Collectors.joining(EXAMPLE_PARAMETER_DELIMITER,
+								CODE_REFERENCE_ITEM_START,
+								CODE_REFERENCE_ITEM_END
+						))
 		);
 	}
 
@@ -311,7 +332,9 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 		}
 		String result = step;
 		for (Map.Entry<String, String> e : example.entrySet()) {
-			result = result.replaceAll(String.format(EXAMPLE_VALUE_PATTERN, e.getKey()), Matcher.quoteReplacement(e.getValue()));
+			result = result.replaceAll(String.format(EXAMPLE_VALUE_PATTERN, e.getKey()),
+					Matcher.quoteReplacement(e.getValue())
+			);
 		}
 		return result;
 	}
@@ -351,7 +374,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 				.collect(Collectors.toList()));
 		usedParams.ifPresent(rq::setParameters);
 		rq.setTestCaseId(ofNullable(getTestCaseId(codeRef,
-				usedParams.map(p -> p.stream().map(ParameterResource::getValue).collect(Collectors.toList())).orElse(null)
+				usedParams.map(p -> p.stream().map(ParameterResource::getValue).collect(Collectors.toList()))
+						.orElse(null)
 		)).map(TestCaseIdEntry::getId).orElse(null));
 		return rq;
 	}
@@ -360,13 +384,16 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * Extension point to customize lifecycle suite (before/after) creation event/request
 	 *
 	 * @param name      a lifecycle suite name
+	 * @param codeRef   the suite code reference
 	 * @param startTime item start time
 	 * @return Request to ReportPortal
 	 */
 	@Nonnull
-	protected StartTestItemRQ buildLifecycleSuiteStartRq(@Nonnull final String name, @Nullable final Date startTime) {
+	protected StartTestItemRQ buildLifecycleSuiteStartRq(@Nonnull final String name, @Nullable String codeRef,
+			@Nullable final Date startTime) {
 		StartTestItemRQ rq = new StartTestItemRQ();
 		rq.setName(name);
+		rq.setCodeRef(codeRef);
 		rq.setStartTime(ofNullable(startTime).orElseGet(() -> Calendar.getInstance().getTime()));
 		rq.setType(ItemType.TEST.name());
 		return rq;
@@ -377,15 +404,16 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 *
 	 * @param type      item type
 	 * @param name      a lifecycle method name
+	 * @param codeRef   method's code reference
 	 * @param startTime item start time
 	 * @return Request to ReportPortal
 	 */
 	@Nonnull
 	protected StartTestItemRQ buildLifecycleMethodStartRq(@Nonnull final ItemType type, @Nonnull final String name,
-			@Nullable final Date startTime) {
+			@Nullable final String codeRef, @Nullable final Date startTime) {
 		StartTestItemRQ rq = new StartTestItemRQ();
 		rq.setName(name);
-		rq.setCodeRef(name);
+		rq.setCodeRef(codeRef);
 		rq.setStartTime(ofNullable(startTime).orElseGet(() -> Calendar.getInstance().getTime()));
 		rq.setType(type.name());
 		return rq;
@@ -464,23 +492,26 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 					Optional.empty() :
 					Optional.of(leafChain.get(leafChain.size() - 1));
 			final TestItemTree.TestItemLeaf parentLeaf = parent.map(Pair::getValue).orElse(null);
-			final Map<TestItemTree.ItemTreeKey, TestItemTree.TestItemLeaf> children = parent.map(p -> p.getValue().getChildItems())
-					.orElseGet(itemTree::getTestItems);
+			final Map<TestItemTree.ItemTreeKey, TestItemTree.TestItemLeaf> children = parent.map(p -> p.getValue()
+					.getChildItems()).orElseGet(itemTree::getTestItems);
 			final String parentCodeRef = parent.map(p -> (String) p.getValue().getAttribute(CODE_REF)).orElse(null);
 			Date itemDate = getItemDate(parent.map(Pair::getValue).orElse(null));
 			switch (itemType) {
 				case STORY:
 					Story story = (Story) entity.get();
 					TestItemTree.ItemTreeKey storyKey = ItemTreeUtils.createKey(story);
-					leafChain.add(ImmutablePair.of(storyKey, children.computeIfAbsent(storyKey, k -> createLeaf(ItemType.STORY,
-							buildStartStoryRq(story, getCodeRef(parentCodeRef, k, ItemType.STORY), itemDate),
-							parentLeaf
-					))));
+					leafChain.add(ImmutablePair.of(storyKey,
+							children.computeIfAbsent(storyKey, k -> createLeaf(ItemType.STORY,
+									buildStartStoryRq(story, getCodeRef(parentCodeRef, k, ItemType.STORY), itemDate),
+									parentLeaf
+							))
+					));
 					break;
 				case SCENARIO:
 					Scenario scenario = (Scenario) entity.get();
 					TestItemTree.ItemTreeKey scenarioKey = ItemTreeUtils.createKey(getScenarioName(scenario));
-					leafChain.add(ImmutablePair.of(scenarioKey, children.computeIfAbsent(scenarioKey, k -> createLeaf(ItemType.SCENARIO,
+					leafChain.add(ImmutablePair.of(scenarioKey, children.computeIfAbsent(scenarioKey, k -> createLeaf(
+							ItemType.SCENARIO,
 							buildStartScenarioRq(scenario, getCodeRef(parentCodeRef, k, ItemType.SCENARIO), itemDate),
 							parentLeaf
 					))));
@@ -500,13 +531,19 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 				case TEST: // type TEST == a lifecycle SUITE
 					String lifecycleSuiteName = (String) entity.get();
 					TestItemTree.ItemTreeKey lifecycleSuiteKey = ItemTreeUtils.createKey(lifecycleSuiteName);
-					leafChain.add(ImmutablePair.of(lifecycleSuiteKey, children.computeIfAbsent(lifecycleSuiteKey,
-							k -> createLeaf(itemType, buildLifecycleSuiteStartRq(lifecycleSuiteName, itemDate), parentLeaf)
-					)));
+					leafChain.add(ImmutablePair.of(lifecycleSuiteKey,
+							children.computeIfAbsent(lifecycleSuiteKey, k -> createLeaf(itemType,
+									buildLifecycleSuiteStartRq(lifecycleSuiteName,
+											getCodeRef(parentCodeRef, k, ItemType.TEST),
+											itemDate
+									),
+									parentLeaf
+							))
+					));
 					break;
 			}
 		}
-		return leafChain.get(leafChain.size() - 1).getValue();
+		return leafChain.isEmpty() ? null : leafChain.get(leafChain.size() - 1).getValue();
 	}
 
 	/**
@@ -559,7 +596,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @param parent a parent test item leaf
 	 * @return the step leaf
 	 */
-	protected TestItemTree.TestItemLeaf startStep(@Nonnull final String name, @Nonnull final TestItemTree.TestItemLeaf parent) {
+	protected TestItemTree.TestItemLeaf startStep(@Nonnull final String name,
+			@Nonnull final TestItemTree.TestItemLeaf parent) {
 		TestItemTree.ItemTreeKey key = ItemTreeUtils.createKey(name);
 		TestItemTree.TestItemLeaf leaf = createLeaf(ItemType.STEP, buildStartStepRq(name,
 				getCodeRef(parent.getAttribute(CODE_REF), key, ItemType.STEP),
@@ -579,10 +617,14 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @return the step leaf
 	 */
 	@Nonnull
-	protected TestItemTree.TestItemLeaf startLifecycleMethod(@Nonnull final String name, @Nonnull final ItemType itemType,
-			@Nonnull final TestItemTree.TestItemLeaf parent) {
+	protected TestItemTree.TestItemLeaf startLifecycleMethod(@Nonnull final String name,
+			@Nonnull final ItemType itemType, @Nonnull final TestItemTree.TestItemLeaf parent) {
 		TestItemTree.ItemTreeKey key = ItemTreeUtils.createKey(name);
-		TestItemTree.TestItemLeaf leaf = createLeaf(itemType, buildLifecycleMethodStartRq(itemType, name, getItemDate(parent)), parent);
+		String codeRef = getCodeRef(parent.getAttribute(CODE_REF), key, itemType);
+		TestItemTree.TestItemLeaf leaf = createLeaf(itemType,
+				buildLifecycleMethodStartRq(itemType, name, codeRef, getItemDate(parent)),
+				parent
+		);
 		parent.getChildItems().put(key, leaf);
 		return leaf;
 	}
@@ -597,8 +639,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 */
 	@SuppressWarnings("unused")
 	@Nonnull
-	protected FinishTestItemRQ buildFinishTestItemRequest(@Nonnull final Maybe<String> id, @Nullable final ItemStatus status,
-			@Nullable Issue issue) {
+	protected FinishTestItemRQ buildFinishTestItemRequest(@Nonnull final Maybe<String> id,
+			@Nullable final ItemStatus status, @Nullable Issue issue) {
 		FinishTestItemRQ rq = new FinishTestItemRQ();
 		rq.setEndTime(Calendar.getInstance().getTime());
 		rq.setStatus(ofNullable(status).map(Enum::name).orElse(null));
@@ -643,7 +685,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @see StatusEvaluation#evaluateStatus(ItemStatus, ItemStatus)
 	 */
 	@Nullable
-	protected ItemStatus evaluateStatus(@Nullable final ItemStatus currentStatus, @Nullable final ItemStatus childStatus) {
+	protected ItemStatus evaluateStatus(@Nullable final ItemStatus currentStatus,
+			@Nullable final ItemStatus childStatus) {
 		return StatusEvaluation.evaluateStatus(currentStatus, childStatus);
 	}
 
@@ -674,7 +717,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @return a {@link SaveLogRQ} supplier {@link Function}
 	 */
 	@Nonnull
-	protected Function<String, SaveLogRQ> getLogSupplier(@Nonnull final LogLevel level, @Nullable final String message) {
+	protected Function<String, SaveLogRQ> getLogSupplier(@Nonnull final LogLevel level,
+			@Nullable final String message) {
 		return itemUuid -> {
 			SaveLogRQ rq = new SaveLogRQ();
 			rq.setItemUuid(itemUuid);
@@ -693,7 +737,9 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @param thrown {@link Throwable} object with details of the failure
 	 */
 	protected void sendStackTraceToRP(@Nonnull Maybe<String> itemId, @Nullable final Throwable thrown) {
-		ofNullable(thrown).ifPresent(t -> ReportPortal.emitLog(itemId, getLogSupplier(LogLevel.ERROR, ExceptionUtils.getStackTrace(t))));
+		ofNullable(thrown).ifPresent(t -> ReportPortal.emitLog(itemId,
+				getLogSupplier(LogLevel.ERROR, ExceptionUtils.getStackTrace(t))
+		));
 	}
 
 	/**
@@ -703,12 +749,15 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @param status a status of the item which will be set
 	 * @param issue  an optional issue which will be set
 	 */
-	protected void finishItem(final @Nonnull Maybe<String> id, final @Nonnull ItemStatus status, @Nullable Issue issue) {
+	protected void finishItem(final @Nonnull Maybe<String> id, final @Nonnull ItemStatus status,
+			@Nullable Issue issue) {
 		FinishTestItemRQ rq = buildFinishTestItemRequest(id, status, issue);
+		//noinspection ReactiveStreamsUnusedPublisher
 		launch.get().finishTestItem(id, rq);
 	}
 
-	private void finishStep(final @Nonnull TestItemTree.TestItemLeaf step, final @Nonnull ItemStatus status, @Nullable Issue issue) {
+	private void finishStep(final @Nonnull TestItemTree.TestItemLeaf step, final @Nonnull ItemStatus status,
+			@Nullable Issue issue) {
 		finishItem(step.getItemId(), status, issue);
 		step.setStatus(status);
 	}
@@ -770,7 +819,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 */
 	@Override
 	public void beforeStory(@Nonnull Story story, boolean givenStory) {
-		currentLifecycleItemType = AFTER_STORIES.equals(story.getName()) ? ItemType.AFTER_SUITE : ItemType.BEFORE_SUITE;
+		currentLifecycleTopItemType = ItemType.AFTER_GROUPS;
+		currentLifecycleItemType = ItemType.BEFORE_SUITE;
 		structure.add(new Entity<>(ItemType.STORY, story));
 	}
 
@@ -810,7 +860,7 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * Finishes scenario in ReportPortal
 	 */
 	@Override
-	public void afterScenario() {
+	public void afterScenario(Timing timing) {
 		TestItemTree.TestItemLeaf previousItem = getLeaf();
 		if (previousItem != null && previousItem.getType() == ItemType.TEST) {
 			evaluateAndFinishLastItem();
@@ -825,13 +875,38 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 * @param step Step to be reported
 	 */
 	@Override
-	public void beforeStep(@Nonnull String step) {
+	public void beforeStep(@Nonnull Step step) {
 		TestItemTree.TestItemLeaf previousItem = getLeaf();
 		if (previousItem != null && previousItem.getType() == ItemType.TEST) {
+			// Finish Before/After methods
 			evaluateAndFinishLastItem();
 		}
+		TestItemTree.TestItemLeaf parentItem = retrieveLeaf();
+		if (parentItem == null) {
+			// Before Stories
+			structure.add(new Entity<>(ItemType.TEST,
+					currentLifecycleTopItemType == null ? BEFORE_STORIES : AFTER_STORIES
+			));
+			if (currentLifecycleTopItemType == null) {
+				currentLifecycleTopItemType = ItemType.BEFORE_GROUPS;
+			}
+		} else if (parentItem.getType() == ItemType.STORY) {
+			// Before Story
+			structure.add(new Entity<>(ItemType.TEST,
+					currentLifecycleItemType == ItemType.BEFORE_SUITE ? BEFORE_STORY : AFTER_STORY
+			));
+		}
+		if (parentItem == null) {
+			ofNullable(retrieveLeaf()).map(i -> startLifecycleMethod(
+					step.getStepAsString(),
+					currentLifecycleTopItemType,
+					i
+			)).ifPresent(stepStack::add);
+			return;
+		}
 		currentLifecycleItemType = ItemType.BEFORE_METHOD;
-		TestItemTree.TestItemLeaf stepLeaf = ofNullable(retrieveLeaf()).map(l -> startStep(step, l)).orElse(null);
+		TestItemTree.TestItemLeaf stepLeaf = ofNullable(retrieveLeaf()).map(l -> startStep(step.getStepAsString(), l))
+				.orElse(null);
 		stepStack.add(stepLeaf);
 		if (stepLeaf != null) {
 			lastStep = stepLeaf;
@@ -846,7 +921,8 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	@Override
 	public void example(Map<String, String> tableRow, int exampleIndex) {
 		TestItemTree.TestItemLeaf previousItem = getLeaf();
-		if (previousItem != null && (previousItem.getType() == ItemType.TEST || previousItem.getType() == ItemType.SUITE)) {
+		if (previousItem != null && (previousItem.getType() == ItemType.TEST
+				|| previousItem.getType() == ItemType.SUITE)) {
 			evaluateAndFinishLastItem();
 		}
 		structure.add(new Entity<>(ItemType.SUITE, tableRow)); // type SUITE is used for Examples
@@ -864,13 +940,24 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 		evaluateAndFinishLastItem();
 	}
 
+	private void finishBeforeAfterSuites(TestItemTree.TestItemLeaf item) {
+		if (item.getType() == ItemType.BEFORE_GROUPS || item.getType() == ItemType.AFTER_GROUPS) {
+			Entity<?> parentSuite = structure.getLast();
+			evaluateAndFinishLastItem(); // Finish parent suite
+			structure.add(parentSuite);
+		}
+	}
+
 	/**
 	 * Finishes step in ReportPortal
 	 */
 	@Override
 	public void successful(String step) {
 		currentLifecycleItemType = ItemType.AFTER_TEST;
-		ofNullable(stepStack.remove()).ifPresent(s -> finishStep(s, ItemStatus.PASSED));
+		ofNullable(stepStack.pollLast()).ifPresent(s -> {
+			finishStep(s, ItemStatus.PASSED);
+			finishBeforeAfterSuites(s);
+		});
 	}
 
 	/**
@@ -881,21 +968,10 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 */
 	@Override
 	public void failed(String step, Throwable cause) {
-		TestItemTree.TestItemLeaf item = retrieveLeaf();
-		boolean isLifecycleMethod = item == null || stepStack.isEmpty();
-		if (isLifecycleMethod) {
-			if (item == null) {
-				// failed @BeforeStories (annotated) methods
-				structure.add(new Entity<>(ItemType.TEST, currentLifecycleItemType == null ? BEFORE_STORIES : AFTER_STORIES));
-			} else if (item.getType() == ItemType.STORY) {
-				// failed @BeforeStory, @BeforeScenario (annotated) methods
-				structure.add(new Entity<>(ItemType.TEST, currentLifecycleItemType == ItemType.BEFORE_SUITE ? BEFORE_STORY : AFTER_STORY));
-			}
-			stepStack.add(ofNullable(retrieveLeaf()).map(i -> startLifecycleMethod(step, currentLifecycleItemType, i)).orElse(null));
-		}
 		ofNullable(stepStack.pollLast()).ifPresent(i -> {
 			sendStackTraceToRP(i.getItemId(), cause);
 			finishStep(i, ItemStatus.FAILED);
+			finishBeforeAfterSuites(i);
 		});
 	}
 
@@ -906,10 +982,10 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 */
 	@Override
 	public void ignorable(String step) {
-		ofNullable(retrieveLeaf()).ifPresent(l -> {
-			TestItemTree.TestItemLeaf leaf = startStep(step, l);
-			createIgnoredSteps(step, leaf);
-			finishStep(leaf, ItemStatus.SKIPPED);
+		ofNullable(stepStack.pollLast()).ifPresent(i -> {
+			createIgnoredSteps(step, i);
+			finishStep(i, ItemStatus.SKIPPED);
+			finishBeforeAfterSuites(i);
 		});
 	}
 
@@ -920,26 +996,24 @@ public abstract class ReportPortalStoryReporter extends NullStoryReporter {
 	 */
 	@Override
 	public void notPerformed(String step) {
-		TestItemTree.TestItemLeaf item = retrieveLeaf();
-		ofNullable(item).ifPresent(i -> {
-			TestItemTree.TestItemLeaf leaf = startStep(step, i);
-			createNotPerformedSteps(step, leaf);
-			finishStep(leaf, ItemStatus.SKIPPED, Launch.NOT_ISSUE);
+		ofNullable(stepStack.pollLast()).ifPresent(i -> {
+			createNotPerformedSteps(step, i);
+			finishStep(i, ItemStatus.SKIPPED, Launch.NOT_ISSUE);
+			finishBeforeAfterSuites(i);
 		});
 	}
 
 	@Override
 	public void pending(String step) {
-		TestItemTree.TestItemLeaf item = retrieveLeaf();
-		ofNullable(item).ifPresent(i -> {
-			TestItemTree.TestItemLeaf leaf = startStep(step, i);
-			createPendingSteps(step, leaf);
-			finishStep(leaf, ItemStatus.SKIPPED);
+		ofNullable(stepStack.pollLast()).ifPresent(i -> {
+			createPendingSteps(step, i);
+			finishStep(i, ItemStatus.SKIPPED);
+			finishBeforeAfterSuites(i);
 		});
 	}
 
 	@Override
-	public void scenarioNotAllowed(Scenario scenario, String filter) {
+	public void scenarioExcluded(Scenario scenario, String filter) {
 		if (null != scenario.getExamplesTable() && scenario.getExamplesTable().getRowCount() > 0) {
 			beforeExamples(scenario.getSteps(), scenario.getExamplesTable());
 			for (int i = 0; i < scenario.getExamplesTable().getRowCount(); i++) {