+++ src/test/java/com/epam/reportportal/jbehave/lifecycle/AfterStepTest.java
@@ -19,7 +19,6 @@ package com.epam.reportportal.jbehave.lifecycle;
 import com.epam.reportportal.jbehave.BaseTest;
 import com.epam.reportportal.jbehave.ReportPortalStepFormat;
 import com.epam.reportportal.jbehave.integration.basic.EmptySteps;
-import com.epam.reportportal.jbehave.integration.lifecycle.AfterScenarioFailedSteps;
 import com.epam.reportportal.listeners.ItemStatus;
 import com.epam.reportportal.listeners.ItemType;
 import com.epam.reportportal.service.ReportPortal;
@@ -35,22 +34,32 @@ import org.mockito.ArgumentCaptor;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.nullValue;
 import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.*;
 
-public class VerifyAfterScenarioAnnotationFailed extends BaseTest {
+public class AfterStepTest extends BaseTest {
 
+	private static final int STEP_NUMBER = 2;
 	private final String storyId = CommonUtils.namedId("story_");
 	private final String scenarioId = CommonUtils.namedId("scenario_");
-	private final String afterStepId = CommonUtils.namedId("after_scenario_step_");
-	private final String stepId = CommonUtils.namedId("step_");
+	private final List<String> afterStepIds = Stream.generate(() -> CommonUtils.namedId("after_step_"))
+			.limit(STEP_NUMBER)
+			.collect(Collectors.toList());
+	private final List<String> stepIds = Stream.generate(() -> CommonUtils.namedId("step_"))
+			.limit(STEP_NUMBER)
+			.collect(Collectors.toList());
 
 	private final List<Pair<String, List<String>>> steps = Collections.singletonList(Pair.of(scenarioId,
-			Arrays.asList(stepId, afterStepId)
+			IntStream.range(0, stepIds.size())
+					.mapToObj(i -> Stream.of(stepIds.get(i), afterStepIds.get(i)))
+					.flatMap(i -> i)
+					.collect(Collectors.toList())
 	));
 
 	private final ReportPortalClient client = mock(ReportPortalClient.class);
@@ -65,60 +74,57 @@ public class VerifyAfterScenarioAnnotationFailed extends BaseTest {
 		mockBatchLogging(client);
 	}
 
-	private static final String STORY_PATH = "stories/NoScenario.story";
-	private static final String DEFAULT_SCENARIO_NAME = "No name";
-	private static final String STEP_NAME = "Given I have empty step";
+	private static final String STORY_PATH = "stories/lifecycle/AfterStep.story";
+	private static final String SCENARIO_NAME = "The scenario";
+	private static final String[] STEP_NAMES = new String[] { "Given I have empty step",
+			"When I have one more empty step" };
+	private static final String LIFECYCLE_STEP_NAME = "Then I have another empty step";
 
 	@Test
-	public void verify_after_scenario_annotation_failed_method_reporting() {
-		run(format, STORY_PATH, new AfterScenarioFailedSteps(), new EmptySteps());
+	public void verify_after_step_lifecycle_step_reporting() {
+		run(format, STORY_PATH, new EmptySteps());
 
 		verify(client).startTestItem(any());
 		ArgumentCaptor<StartTestItemRQ> startCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
 		verify(client).startTestItem(same(storyId), startCaptor.capture());
-		verify(client, times(2)).startTestItem(same(scenarioId), startCaptor.capture());
+		verify(client, times(STEP_NUMBER * 2)).startTestItem(same(scenarioId), startCaptor.capture());
 
 		// Start items verification
 		List<StartTestItemRQ> startItems = startCaptor.getAllValues();
-		String scenarioCodeRef = STORY_PATH + String.format("/[SCENARIO:%s]", DEFAULT_SCENARIO_NAME);
 		StartTestItemRQ scenarioStart = startItems.get(0);
-		assertThat(scenarioStart.getName(), equalTo(DEFAULT_SCENARIO_NAME));
+		String scenarioCodeRef = STORY_PATH + String.format(SCENARIO_PATTERN, SCENARIO_NAME);
+		assertThat(scenarioStart.getName(), equalTo(SCENARIO_NAME));
 		assertThat(scenarioStart.getCodeRef(), equalTo(scenarioCodeRef));
 		assertThat(scenarioStart.getType(), equalTo(ItemType.SCENARIO.name()));
 
-		StartTestItemRQ step = startItems.get(1);
-		String stepCodeRef = scenarioCodeRef + String.format("/[STEP:%s]", STEP_NAME);
-		assertThat(step.getName(), equalTo(STEP_NAME));
-		assertThat(step.getCodeRef(), equalTo(stepCodeRef));
-		assertThat(step.getType(), equalTo(ItemType.STEP.name()));
+		List<StartTestItemRQ> steps = Arrays.asList(startItems.get(1), startItems.get(3));
+		IntStream.range(0, steps.size()).forEach(i -> {
+			StartTestItemRQ step = steps.get(i);
+			String stepCodeRef = scenarioCodeRef + String.format(STEP_PATTERN, STEP_NAMES[i]);
+			assertThat(step.getName(), equalTo(STEP_NAMES[i]));
+			assertThat(step.getCodeRef(), equalTo(stepCodeRef));
+			assertThat(step.getType(), equalTo(ItemType.STEP.name()));
+		});
 
-		StartTestItemRQ afterStep = startItems.get(2);
-		String afterStepCodeRef = AfterScenarioFailedSteps.class.getCanonicalName() + ".afterScenarioFailed()";
-		assertThat(afterStep.getName(), equalTo(afterStepCodeRef));
-		assertThat(afterStep.getCodeRef(), equalTo(afterStepCodeRef));
-		assertThat(afterStep.getType(), equalTo(ItemType.AFTER_TEST.name()));
+		List<StartTestItemRQ> afterStarts = Arrays.asList(startItems.get(2), startItems.get(4));
+		IntStream.range(0, afterStarts.size()).forEach(i -> {
+			StartTestItemRQ beforeStart = afterStarts.get(i);
+			String lifecycleCodeRef = scenarioCodeRef + String.format(STEP_PATTERN, LIFECYCLE_STEP_NAME);
+			assertThat(beforeStart.getName(), equalTo(LIFECYCLE_STEP_NAME));
+			assertThat(beforeStart.getCodeRef(), equalTo(lifecycleCodeRef));
+			assertThat(beforeStart.getType(), equalTo(ItemType.STEP.name()));
+		});
 
 		// Finish items verification
 		ArgumentCaptor<FinishTestItemRQ> finishStepCaptor = ArgumentCaptor.forClass(FinishTestItemRQ.class);
-		verify(client).finishTestItem(same(stepId), finishStepCaptor.capture());
-		verify(client).finishTestItem(same(afterStepId), finishStepCaptor.capture());
+		verify(client).finishTestItem(same(stepIds.get(0)), finishStepCaptor.capture());
+		verify(client).finishTestItem(same(afterStepIds.get(0)), finishStepCaptor.capture());
+		verify(client).finishTestItem(same(stepIds.get(1)), finishStepCaptor.capture());
+		verify(client).finishTestItem(same(afterStepIds.get(1)), finishStepCaptor.capture());
 		verify(client).finishTestItem(same(scenarioId), finishStepCaptor.capture());
 		verify(client).finishTestItem(same(storyId), finishStepCaptor.capture());
 
 		List<FinishTestItemRQ> finishItems = finishStepCaptor.getAllValues();
-
-		FinishTestItemRQ stepFinish = finishItems.get(0);
-		assertThat(stepFinish.getStatus(), equalTo(ItemStatus.PASSED.name()));
-		assertThat(stepFinish.getIssue(), nullValue());
-
-		FinishTestItemRQ beforeStepFinish = finishItems.get(1);
-		assertThat(beforeStepFinish.getStatus(), equalTo(ItemStatus.FAILED.name()));
-		assertThat(beforeStepFinish.getIssue(), nullValue());
-
-		FinishTestItemRQ scenarioFinish = finishItems.get(2);
-		assertThat(scenarioFinish.getStatus(), equalTo(ItemStatus.FAILED.name()));
-
-		FinishTestItemRQ storyFinish = finishItems.get(3);
-		assertThat(storyFinish.getStatus(), equalTo(ItemStatus.FAILED.name()));
+		finishItems.forEach(i -> assertThat(i.getStatus(), equalTo(ItemStatus.PASSED.name())));
 	}
 }