+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## [Unreleased]
-### Changed
-- Test new actions
+### Added
+- JBehave version 5 support, by @HardNorth
+### Removed
+- JBehave version 4.8 support, by @HardNorth
 
 ## [5.1.3]
 ### Changed